+++ src/Resource/Order/Carriers/Allegro.php
@@ -6,7 +6,6 @@ namespace Imper86\PhpAllegroApi\Resource\Order\Carriers;
 
 use Imper86\PhpAllegroApi\Resource\AbstractResource;
 use Imper86\PhpAllegroApi\Resource\Order\Carriers\Allegro\Points;
-use Imper86\PhpAllegroApi\Resource\Order\Carriers\Allegro\Tracking;
 
 /**
  * @method Points points()
