+++ src/Resource/Order/Carriers/Tracking.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Imper86\PhpAllegroApi\Resource\Order\Carriers\Allegro;
+namespace Imper86\PhpAllegroApi\Resource\Order\Carriers;
 
 use Imper86\PhpAllegroApi\Enum\ContentType;
 use Imper86\PhpAllegroApi\Resource\AbstractResource;
@@ -10,14 +10,15 @@ use Psr\Http\Message\ResponseInterface;
 class Tracking extends AbstractResource
 {
     /**
+     * @param string $carrierId
      * @param string[] $waybills
      * @return ResponseInterface
      * @throws ClientExceptionInterface
      */
-    public function get(array $waybills): ResponseInterface
+    public function get(string $carrierId, array $waybills): ResponseInterface
     {
         return $this->apiGet(
-            '/order/carriers/ALLEGRO/tracking',
+            "/order/carriers/{$carrierId}/tracking",
             ['waybill' => $waybills],
             ContentType::VND_PUBLIC_V1
         );
